Rename `StdLoggerBackend` to `JulLoggerBackend`, and move it to `backend.jul` package
